Move each test to its proper package
